BINDER_WRITE_READ initalised as i32 and again converted it to i32, I know it's weird but I guess that's RUST

+++ server/mirrox_server_jar/rust_binder/src/binder_transaction.rs
@@ -7,7 +7,7 @@ use nix::unistd::geteuid;
 
 // Required ioctl number for BINDER_WRITE_READ
 // const BINDER_WRITE_READ: u64 = 0xC0306201;
-const BINDER_WRITE_READ: u64 = 0xC0306201u32 as i32;
+const BINDER_WRITE_READ: i32 = 0xC0306201u32 as i32;
 const BR_TRANSACTION: u32 = 0x1c;
 const BR_REPLY: u32 = 0x1f;
